+++ src/sections/scheme/view/scheme-list-view.jsx
@@ -115,6 +115,10 @@ export default function SchemeListView() {
   }, []);
 
   const handleDelete = async (id) => {
+    if (!getResponsibilityValue('delete_scheme', configs, user)) {
+      enqueueSnackbar('You do not have permission to delete.', { variant: 'error' });
+      return;
+    }
     try {
       const res = await axios.delete(`${import.meta.env.VITE_BASE_URL}/${user?.company}/scheme`, {
         data: { ids: id },
